fix(B-464): rename session_feedback schema file
